+++ CHANGELOG.md
@@ -6,11 +6,19 @@ SPDX-License-Identifier: MIT
 
 # Changelog
 
-## tol-sdk `2.7.0`
+## tol-sdk `2.8.0`
 
-16-09-2026
+21-09-2026
 
 - Added OpenCitations lookup and filtering by DOI and PubMed IDs
+
+## tol-sdk `2.7.0`
+
+17-09-2026
+
+- Add quasar() source
+- Add support for direct and via-api versions of datasources
+- Add assembly_span from GoaT
 
 ## tol-sdk `2.6.0`